+++ src/onetick/py/sources/__init__.py
@@ -18,3 +18,4 @@ from .split_query_output_by_symbol import SplitQueryOutputBySymbol, by_symbol
 from .symbology_mapping import SymbologyMapping
 from .symbols import Symbols
 from .pit import PointInTime
+from .ref_data import RefData